Creating Controller and service

+++ src/main/java/com/example/usermicroservice/Controller/UserController.java
@@ -1,8 +1,10 @@
 package com.example.usermicroservice.Controller;
 
 import com.example.usermicroservice.Entity.User;
+import com.example.usermicroservice.Service.AdminService;
 import com.example.usermicroservice.Service.UserService;
 import com.example.usermicroservice.repository.UserRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import lombok.AllArgsConstructor;
@@ -19,34 +21,40 @@ public class UserController {
 
     private UserService userService;
 
+    private AdminService adminService;;
+
+
     @GetMapping("/get")
-    public ResponseEntity<List<User>> getAllBooks() {
+    public ResponseEntity<List<User>> getAllUser() {
         List<User> users = userService.getAllUsers();
         return ResponseEntity.ok(users);
     }
 
     @GetMapping("/get/{id}")
-    public ResponseEntity<User> getBookById(@PathVariable long id) {
+    public ResponseEntity<User> getUserById(@PathVariable long id) {
         Optional<User> user =  userService.getUserById(id);
         return user.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    @PostMapping("/add")
-    public ResponseEntity<User> addBook(@RequestBody User User) {
+    @PostMapping("/addUser")
+    public ResponseEntity<User> addUser(@RequestBody User User) {
         User user = userService.createUser(User);
         return ResponseEntity.ok(user);
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<User> deleteBook(@PathVariable long id) {
+    public ResponseEntity<String> deleteUser(@PathVariable long id) {
+        if (!adminService.CheckIfAdmin(id)){
+            return new ResponseEntity<>("You're not an admin", HttpStatus.NOT_FOUND);
+        }
         if (userService.deleteUserById(id)){
-            return ResponseEntity.noContent().build();
+            return ResponseEntity.ok("User deleted successfully");
         }
         return ResponseEntity.notFound().build();
     }
 
     @PutMapping("/edit/{id}")
-    public ResponseEntity<User> editBook(@PathVariable long id, @RequestBody User User) {
+    public ResponseEntity<User> editUser(@PathVariable long id, @RequestBody User User) {
         boolean isUpdated = userService.updateUserById(id, User);
         if (isUpdated) {
             return ResponseEntity.ok(User);
